+++ components/restreamer/client/cypress/integration/add_input_pull.js
@@ -19,7 +19,7 @@ describe('ADD PULL INPUT', () => {
   });
 
   it('Set rtmp://', () => {
-    const urlToPaste = 'rtmp://' + Cypress.env('host') + '/en/origin';
+    const urlToPaste = 'rtmp://' + Cypress.env('host') + '/en/primary';
     cy.get('[placeholder="rtmp://..."]')
       .invoke('val', urlToPaste)
       .trigger('input');
@@ -31,9 +31,9 @@ describe('ADD PULL INPUT', () => {
   });
 
   it('Assert', () => {
-    cy.get("span:contains('/en/origin'):last").should(
+    cy.get("span:contains('/en/primary'):last").should(
       'have.text',
-      'rtmp://' + Cypress.env('host') + '/en/origin'
+      'rtmp://' + Cypress.env('host') + '/en/primary'
     );
   });
 });