+++ components/restreamer/client/cypress/integration/check_streaming_state.js
@@ -3,7 +3,7 @@ describe('CHECK STREAMING STATE', () => {
     cy.visit('/');
     cy.deleteAllInputs();
     cy.importJsonConf(Cypress.env('host'));
-    cy.runTestStream('rtmp://' + Cypress.env('host') + '/en/origin');
+    cy.runTestStream('rtmp://' + Cypress.env('host') + '/en/primary');
   });
 
   it('1 Assert Start All', () => {